// src/types/Category.ts
type Category =
    | "beauty"
    | "books"
    | "books & reference"
    | "business"
    | "cars"
    | "dating"
    | "design"
    | "developer"
    | "developer tools"
    | "development"
    | "education"
    | "entertainment"
    | "events"
    | "fashion"
    | "finance"
    | "fitness"
    | "food"
    | "fundraising"
    | "games"
    | "government"
    | "graphics"
    | "graphics & design"
    | "health"
    | "health & fitness"
    | "kids"
    | "lifestyle"
    | "magazines"
    | "medical"
    | "multimedia"
    | "multimedia design"
    | "music"
    | "navigation"
    | "network"
    | "networking"
    | "news"
    | "parenting"
    | "personalization"
    | "pets"
    | "photo"
    | "photo & video"
    | "politics"
    | "productivity"
    | "reference"
    | "security"
    | "shopping"
    | "social"
    | "social networking"
    | "sports"
    | "transportation"
    | "travel"
    | "utilities"
    | "video"
    | "weather"

export default Category;